fix: extract leaderboard zod schemas to shared file to prevent hydration error
The leaderboard client imported zod schemas from the API route files, which
also import the DB client at the top level. When Vite bundled for the browser,
it followed the full import chain and the LibsqlDialect constructor threw
because DB env vars don't exist on the client.

Moves batting and bowling zod schemas into a new `src/lib/leaderboard-schemas.ts`
file with zero server-side imports. The API routes and client now both import
from this shared file.

Closes #209
